perf: wait for initial load in benchmarks
